Main/SystemTableBuilder.cpp: size_t -> uint32_t warning. jiraref:VBP-1171
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@106603 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-client/SystemTableBuilder.cpp
@@ -517,8 +517,8 @@ int SystemTableBuilderAcpi::buildMadt(RTVFSIOSTREAM hVfsIos, size_t *pcbMadt)
 
 int SystemTableBuilderAcpi::buildMcfg(RTVFSIOSTREAM hVfsIos, size_t *pcbMcfg)
 {
-    uint8_t abMcfg[_1K];
-    size_t cbMcfg = 0;
+    uint8_t  abMcfg[_1K];
+    uint32_t cbMcfg = 0;
 
     RT_ZERO(abMcfg);
